Fix predicate

+++ src/lib/find_load_command_by_command.c
@@ -9,16 +9,18 @@ struct load_command	*find_load_command_by_command(
 	t_mach_o *mach_o,
 	uint32_t cmd)
 {
-	uint32_t             i;
-	uint32_t             ncmds;
+	size_t             i;
+	size_t             ncmds;
 	struct load_command  *cursor_load_command;
 
-  // When the load_command or the magick number has not been set, return NULL
+  	// When the load_command or the magick number has not been set, return NULL
 	if (NULL == mach_o->load_commands
 	   || (NULL == mach_o->mh && NULL == mach_o->mh_64))
 		return (NULL);
 
   i = 0;
+
+	// printf("AAAAAAAAA\n");
 
   // Read ncmds from 32 or 64 bits structure, according to mh
   ncmds = mach_o->mh ? mach_o->mh->ncmds : mach_o->mh_64->ncmds;
@@ -36,11 +38,11 @@ struct load_command	*find_load_command_by_command(
 			|| (LC_REQ_DYLD | cmd) == cursor_load_command->cmd)
 			return (cursor_load_command);
 
-  	cursor_load_command =
+  		cursor_load_command =
 			(struct load_command *)(void *)((uint8_t*)cursor_load_command
 			+ cursor_load_command->cmdsize);
 
-  	i++;
+  		i++;
 	}
 
   return (NULL);
